refactor: migrate fds pos and evlib into desc
We have to do several refactors to prepare for the ability for client connections
to have multiple wsis attempting the same connection via different paths /
dest addresses.

This starts to prepare for that by expending wsi->desc from being a simple
union to being a struct containing the union, moving the position_in_fds_table from
the wsi into the desc object.

I started off intending to take this further and have the desc usable as a
standalone object on the fds list, but considering event libs the amount of
churn will be much too huge.  So wsis remain the unit of currency for
connections.

As it is, this should be a NOP.