cmake: do not allow netlink support when external poll support is enabled.
Netlink fds are not passed to the user callback for external loop adoption, because vhost and its protocols are not yet in place when rops_pt_init_destroy_netlink() gets called during context creation. As a consequence, netlink fds are not polled, leading to outdated/wrong routing table entries in the library.
A complex rework would be required to support this, so for now disable netlink when external poll is needed.