client-http: improve exit paths
The handling of return codes in client-http generally can do with some
improvement.  Basically it should be OK (0) or the wsi should be closed
synchronously in the function and return nonzero.

Re-use LWS_HPI_RET_WSI_ALREADY_DIED as the closed return to make the
meaning unambiguous.

Add some extra use of bail3 exit path to cleanly return the closed
value with cce for some extra cases.