ss: bring context extensions out in tls vhosts
SS creates vhosts for the life of the client tls pieces the vhost brings in, so it
can get away with dynamically having the necessary support for the tls
pieces in heap only when they are needed.  That's very useful on eg ESP32
that wants to make random internet connections since it can magic up the
vhost with the right tls context using JIT trust and remove it when it's no
longer needed.

Until now it didn't take care about including the context's extension list
on those vhosts, this enables the ss to have access to that.  Specifically
the ss binance example now negotiates pm_deflate with this.