cmake: set minimum to 3.10
CMake is continually deprecating old cmake, such that if we don't try to
keep up with its minimum version there are warnings or eventually failures
to build.  Windows vcpkg has latest 4.03 which prints warning messages if
your CMakeLists.txt has a VERSION < 3.10.  So let's upgrade all the
CMakeLists.txt to that as the minimum.