NN-2025-0102: UAF depending on upgrade allowed
This document contains sensitive information collected during our
security research activities related with the Libwebsockets library
maintained by Andy Green (warmcat).

+-------------------------------------------------------------------------------------------------------+
| Report information                                                                                    |
+:===================================:+:===============================================================:+
| Vendor                              | warmcat                                                         |
+-------------------------------------+-----------------------------------------------------------------+
| Vendor URL                          | https://libwebsockets.org/git/libwebsockets                     |
+-------------------------------------+-----------------------------------------------------------------+
| Affected component                  | libwebsockets                                                   |
+-------------------------------------+-----------------------------------------------------------------+
| Affected version                    | 4.4                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Vulnerability                       | CWE-416: Use After Free                                         |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Base Score       | 6.0                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Vector           | CVSS:4.0/AV:N/AC:L/AT:P/PR:N/UI:P/VC:N/VI:N/VA:H/SC:N/SI:N/SA:N |
+-------------------------------------+-----------------------------------------------------------------+

+-----------------------------------------------------------------------------+
| Security Researcher(s)                                                      |
+:===================================:+:=====================================:+
| Name                                | **Email address**                     |
+-------------------------------------+---------------------------------------+
| Raffaele Bova                       | labs-advisory@nozominetworks.com      |
+-------------------------------------+---------------------------------------+

Libwebsockes is a C library that provides client and server
implementation for various protocols (e.g., HTTP, websockets, MQTT) and
more.

Nozomi Networks Lab discovered a "CWE-416: Use After Free" in the latest
software version of libwebsockets, specifically in the WebSocket server
implementation.

Depending on the use of the API, the vulnerability may allow an attacker
to read or write data, that could cause a loss of integrity or
availability.

The issue is caused by the `lws_handshake_protocol` function, specifically
when the upgrade header is not valid, the function calls
`lws_http_transaction_completed`, which frees some of the data in the wsi
structure, then it calls `user_callback_handle_rxflow` passing the up
pointer and uses it on following strcasecmp calls.

From our understanding, for this vulnerability to have a meaningful
impact, a user that implements the Websocket server, must provide a user
callback function which is going to handle
`LWS_CALLBACK_HTTP_CONFIRM_UPGRADE`, while ignoring the length and doing
operations on the up pointer.

It is possible to compile the minimal websocket server using address
sanitizer, to quickly verify the use after free.

From our understanding of the code, if the upgrade header does not match
the intended contents, then the code after the if statement when
`lws_http_transaction_completed` is called, should not be executed, thus
simply enclosing all that code in the else branch solves the issue.