NN-2025-0103: ADNS crafted response overflow
This document contains sensitive information collected during our
security research activities related with the Libwebsockets library made
by Andy Green (warmcat).

+-------------------------------------------------------------------------------------------------------+
| Report information                                                                                    |
+:===================================:+:===============================================================:+
| Vendor                              | warmcat                                                         |
+-------------------------------------+-----------------------------------------------------------------+
| Vendor URL                          | https://libwebsockets.org/git/libwebsockets                     |
+-------------------------------------+-----------------------------------------------------------------+
| Affected component                  | Ecostruxure Automation Expert                                   |
+-------------------------------------+-----------------------------------------------------------------+
| Affected version                    | 4.4                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Vulnerability                       | CWE-121: Stack-based Buffer Overflow                            |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Base Score       | 7.5                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Vector           | CVSS:4.0/AV:A/AC:L/AT:P/PR:N/UI:P/VC:H/VI:H/VA:H/SC:N/SI:N/SA:N |
+-------------------------------------+-----------------------------------------------------------------+

+-----------------------------------------------------------------------------+
| Security Researcher(s)                                                      |
+:===================================:+:=====================================:+
| Name                                | **Email address**                     |
+-------------------------------------+---------------------------------------+
| Raffaele Bova                       | labs-advisory@nozominetworks.com      |
+-------------------------------------+---------------------------------------+

**\**

Libwebsockes is a C library that provides client and server
implementation for various protocols (e.g., HTTP, websockets, MQTT) and
more.

Nozomi Networks Lab discovered a "CWE-121: Stack-based Buffer Overflow"
in the latest software version of libwebsockets, specifically in the
async-dns component.

The vulnerability allows an attacker that can inspect DNS requests made
by the victim (e.g. being in the same wireless network) to forge a DNS
response packet that overflows the stack and may lead to arbitrary code
execution (depending on the platform and compiler options).

The issue resides in `lws_adns_parse_label` function in
`lib/system/async-dns/async-dns-parse.c`; this function iteratively parses
a label however it does not correctly check the number of bytes written
in the destination buffer.

Specifically, the size of the dest output buffer is specified in the `dl`
argument, however during the read of each substring of the label only
the length of the current substring of the label is accounted for not
overflowing the destination buffer, but previous reads are not accounted
for.

This means that a label of arbitrary size and content can be supplied
and is copied onto the stack, however it must be split into substrings
of size less than `dl`.

To trigger the vulnerability an attacker must be able to sniff the DNS
request packet to send a response with a matching identifier, otherwise
the implantation correctly ignores the response.

We have provided a harness for testing, for ease of use copy the harness
in a subdirectory, for example in minimal-examples-lowlevel/api-tests/,
and build it

```
cmake -B build -DLWS_WITH_SYS_ASYNC_DNS=1 -DLWS_WITH_SSL=0
-DCMAKE_C_FLAGS="-fsanitize=address" . && make -C build lws-test-async-dns
```

Then it can be run `./build/bin/lws-test-async-dns < poc_stackbof`

![Address sanitizer report of stack buffer overflow](./NN-2025-0103_image.png)

We suggest keeping track of the number of bytes currently written on the
dest buffer, this could be done by saving the original dest pointer,
decrementing dl on each substring memcpy, or using an auxiliary
variable.