NN-2025-0104: UPNG out of bounds read
This document contains sensitive information collected during our
security research activities related with the Libwebsockets library made
by Andy Green (warmcat).

+-------------------------------------------------------------------------------------------------------+
| Report information                                                                                    |
+:===================================:+:===============================================================:+
| Vendor                              | warmcat                                                         |
+-------------------------------------+-----------------------------------------------------------------+
| Vendor URL                          | https://libwebsockets.org/git/libwebsockets                     |
+-------------------------------------+-----------------------------------------------------------------+
| Affected component                  | libwebsockets                                                   |
+-------------------------------------+-----------------------------------------------------------------+
| Affected version                    | 4.4                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Vulnerability                       | CWE-125: Out-of-bounds Read                                     |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Base Score       | 6.0                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Vector           | CVSS:4.0/AV:N/AC:L/AT:P/PR:N/UI:P/VC:N/VI:N/VA:H/SC:N/SI:N/SA:N |
+-------------------------------------+-----------------------------------------------------------------+

+-----------------------------------------------------------------------------+
| Security Researcher(s)                                                      |
+:===================================:+:=====================================:+
| Name                                | **Email address**                     |
+-------------------------------------+---------------------------------------+
| Raffaele Bova                       | labs-advisory@nozominetworks.com      |
+-------------------------------------+---------------------------------------+

Libwebsockes is a C library that provides client and server
implementation for various protocols (e.g., HTTP, websockets, MQTT) and
more.

Nozomi Networks Lab discovered a "CWE-125: Out-of-bounds Read" in the
latest software version of libwebsockets, specifically in the PNG
parsing.

Depending on the use of the API, the vulnerability allows an attacker
that can provide a PNG to the software, (e.g., victim visits an attacker
controlled server using the http client in an implementation that uses
the display stack and renders the PNG), could cause a segmentation fault
or crash by making the program read past the heap allocated buffer.

The issue is caused by the `lws_upng_emit_next_line` function,
specifically in the branch when `uf->padded` is true, in the loop where
it reads from the sliding window, ibp is incremented but never reset,
making it possible to read past the buffer, and if the input file is big
enough, reading past the currently mapped heap memory causing a crash.

The way to trigger this issue is to construct a big enough PNG file that
has a width of 1, bit depth of 1 and a very big height.

![Address sanitizer report on out of bounds read](./NN-2025-0104_image.png)

We have provided all the necessary files to reproduce the issue and a
patch that resolves the problem.

From our understanding of the code, it could be sufficient to fix the
array access to the `uf->in` buffer by doing it modulo `u->inf.info_size`,
we don't know whether that is the intended approach, however it resolves
the issue.