NN-2025-0105: UPNG out of bounds write
This document contains sensitive information collected during our
security research activities related with the Libwebsockets library
maintained by Andy Green (warmcat).

+-------------------------------------------------------------------------------------------------------+
| Report information                                                                                    |
+:===================================:+:===============================================================:+
| Vendor                              |   warmcat                                                       |
+-------------------------------------+-----------------------------------------------------------------+
| Vendor URL                          | https://libwebsockets.org/git/libwebsockets                     |
+-------------------------------------+-----------------------------------------------------------------+
| Affected component                  | libwebsockets                                                   |
+-------------------------------------+-----------------------------------------------------------------+
| Affected version                    | 4.4                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Vulnerability                       | CWE-787: Out-of-bounds Write                                    |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Base Score       | 6.0                                                             |
+-------------------------------------+-----------------------------------------------------------------+
| Proposed CVSS v3.1 Vector           | CVSS:4.0/AV:N/AC:L/AT:P/PR:N/UI:P/VC:N/VI:N/VA:H/SC:N/SI:N/SA:N |
+-------------------------------------+-----------------------------------------------------------------+

+-----------------------------------------------------------------------------+
| Security Researcher(s)                                                      |
+:===================================:+:=====================================:+
| Name                                | **Email address**                     |
+-------------------------------------+---------------------------------------+
| Raffaele Bova                       | labs-advisory@nozominetworks.com      |
+-------------------------------------+---------------------------------------+

Libwebsockets is a C library that provides client and server
implementation for various protocols (e.g., HTTP, websockets, MQTT) and
more.

Nozomi Networks Lab discovered a "CWE-787: Out-of-bounds Write" in the
latest software version of libwebsockets, specifically in the upng
component.

The vulnerability allows an attacker to write past the bounds of a heap
allocated buffer, that may lead to a crash.

The issue is caused by the initialization of `bypl` struct member,
whenever the multiplication of width by bypp overflows, resulting in
`bypl` being 0, later on the buffer `u->inf.out` will only be allocated a
buffer of size `u->inf.info_size` and the buffer `u->u.lines` will then
point right after the end of the allocated buffer, thus any subsequent
operation on the lines buffer will write or read past heap allocated
memory.

This bug may lead to a crash or other undesirable effects in the
`unfilter_scanline` function, as the recon buffer directly points inside
the lines buffer.

![Address sanitizer report](./NN-2025-0105_image.png)

We have provided a patch to add a sanity check for the maximum possible
value of width, making sure that the multiplication will not overflow,
as bypp is derived by bpp, and width is also multiplied by bpp, a check
that `width * bpp` will not overflow is needed.