fix(build): build as meson subproject
This add_custom_command declares three OUTPUTs. Meson sees this list and assumes they are all files it needs to track. However, the command cmake -E copy_directory creates a directory named libwebsockets.

When Meson tries to process the outputs of this custom target, it gets to the libwebsockets entry and tries to copy it using a file copy operation, which results in an IsADirectoryError.

The remaining outputs are:

- `${PROJECT_BINARY_DIR}/include/lws_config.h`
- `${PROJECT_BINARY_DIR}/include/libwebsockets.h`

As far as I was able to test, normal cmake build still works successfully after this change.