api-test: lws_struct-json: fixes and additional test
We popped in the lws_struct callback on ARRAY_END and then popped again in
the lejp handler that called it... that caused entries in the JSON after
the array to not be handled correctly because our stack position was wrong
after finishing the array.  Whereas the "same" JSON with the entries done before
the array would be correctly handled.

This fixes the cause and adds the previously working and broken JSON with
the lws_struct pieces to confirm it's fixed.