tls: schannel-phase-2
Let's map the gencrypto stuff on to schannel first.  We can be sure the
build selection pieces are right and test the schannel version of it with
the already-existing api-test-gencrypto.

GCM is broken on schannel, some hours with Gemini couldn't figure out
what it's supposed to do to work so we left it and moved on.

Co-developed-by: Gemini 3.0 Pro